# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API for RockyDex Mobile App, supporting user authentication, favorites, and reading history synchronization.",
    "title": "RockyDex API",
    "contact": {},
    "version": "1.1.2"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Creates a new user account with email, name, and password.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Register a new user",
        "parameters": [
          {
            "description": "Registration details",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.RegisterInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/dto.AuthResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "409": {
            "description": "Conflict - Email already registered",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates a user with email and password, returning a JWT token.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "User Login",
        "parameters": [
          {
            "description": "Login Credentials",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.LoginInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.AuthResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "description": "Authenticates a user using a Google Identity Token. Registers them if not already registered.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Google Authentication Login",
        "parameters": [
          {
            "description": "Google ID Token",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.GoogleLoginInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.AuthResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Generates a 6-digit password reset PIN for the given email. Logs and returns the code (dev-only fallback).",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Request Password Reset Code",
        "parameters": [
          {
            "description": "User Email",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.ForgotPasswordInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset PIN requested",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Resets the user's password using the 6-digit token code received.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Reset User Password",
        "parameters": [
          {
            "description": "Reset details (Email, Token, and New Password)",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.ResetPasswordInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request / Invalid token / Expired",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "dto.RegisterInput": {
      "type": "object",
      "required": [
        "email",
        "password",
        "name"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "dto.LoginInput": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "dto.GoogleLoginInput": {
      "type": "object",
      "required": [
        "id_token"
      ],
      "properties": {
        "id_token": {
          "type": "string"
        }
      }
    },
    "dto.ForgotPasswordInput": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "dto.ResetPasswordInput": {
      "type": "object",
      "required": [
        "email",
        "token",
        "new_password"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "new_password": {
          "type": "string"
        }
      }
    },
    "dto.AuthResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/domain.User"
        }
      }
    },
...
    "domain.User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "profile": {
          "$ref": "#/definitions/domain.Profile"
        }
      }
    },
    "domain.Profile": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string"
        },
        "avatar_url": {
          "type": "string"
        },
        "theme_preference": {
          "type": "string"
        },
        "reading_layout": {
          "type": "string"
        },
        "reading_brightness": {
          "type": "number"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
  }
}